requirements.txt
streamlit
pandas
numpy
scikit-learn
pdfplumber
python-docx
plotly